Started implementation of automated build tool processing
