+++ src/R136.BuildTool/Tasks/Entity.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace R136.BuildTool.Tasks
+{
+	enum Entity
+	{
+		Animates,
+		Commands,
+		Items,
+		Properties,
+		Rooms,
+		Texts
+	}
+}
