+++ src/R136.BuildTool/Tasks/FileNames.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+
+namespace R136.BuildTool.Tasks
+{
+	class FileNames
+	{
+		public string? Input { get; set; }
+		public string? Output { get; set; }
+	}
+}
