+++ src/R136.Shell/appsettings.json
@@ -7,14 +7,14 @@
 			"TitleText": "** Mission code: R136 **",
 			"LanguageSwitchText": "Ok, we'll continue in English.",
 			"ProceedText": "[Press any key]",
-			"LanguageSwitchInstructionText": "To change language, type lang <code>. Available codes are {codes}"
+			"LanguageSwitchInstructionText": "To change language, type lang \\<code\\>. Available codes are {codes}."
 		},
 
 		"nl-NL": {
 			"TitleText": "** Missiecode: R136 **",
 			"LanguageSwitchText": "Ok, we gaan in het Nederlands verder.",
 			"ProceedText": "[Druk op een toets]",
-			"LanguageSwitchInstructionText": "Om de taal te veranderen, type lang <code>. Beschikbare codes zijn {codes}"
+			"LanguageSwitchInstructionText": "Om de taal te veranderen, type lang \\<code\\>. Beschikbare codes zijn {codes}."
 		}
 	},
